+++ shared_engine/sys/sys_in_joystick.cpp
@@ -59,7 +59,7 @@ static CEqGameControllerSDL s_controllers[MAX_CONTROLLERS];
 
 void CEqGameControllerSDL::Init()
 {
-	long mappingsSize = 0;
+	int mappingsSize = 0;
 	const char* mappingsBuf = (const char*)g_fileSystem->GetFileBuffer(CONTROLLER_DB_FILENAME, &mappingsSize);
 	if (mappingsBuf)
 	{
